% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esmax_clusters_small}
\alias{esmax_clusters_small}
\title{Sctype esmax clusters}
\format{
A character vector of cluster assignments for the 1000 cells in esmax_scores_small.
They are in the same order as the rows of esmax_scores_small.
}
\usage{
esmax_clusters_small
}
\description{
A character vector of cluster assignments for the 1000 cells in esmax_scores_small.
}
\keyword{datasets}
