% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esmax_scores_small}
\alias{esmax_scores_small}
\title{Sctype esmax scores}
\format{
A matrix with 1000 rows and 5 columns. Rownames are like this '6_GTTTGTAAGGTGTTAC-1'. The colnames are celltypes. The values are Sctype scores
}
\usage{
esmax_scores_small
}
\description{
A small (1000 rows) esmax scores table taken randomly from the larger 36K original scores table.
}
\keyword{datasets}
