% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_clusterlabel.R
\name{expand_clusterlabel}
\alias{expand_clusterlabel}
\title{Expand Group Labels to Observation Level}
\usage{
expand_clusterlabel(scores, clusters)
}
\arguments{
\item{scores}{data.frame containing at least two columns:
'obs_id' and 'labels' (rows represent groups).}

\item{clusters}{A vector that specifies the group membership for each observation. Must be same length as the number of rows in \code{scores}.}
}
\value{
A vector of the same length as \code{clusters} containing the expanded labels.
}
\description{
Purpose: You have a scores table at the group level. However, you have a
table of metadata at the observation level with a column specifying the group. You
want to add a column of labels to your observation-level metadata. Therefore this function
will expand the group-level labels to a vector that you can add directly to your
observation-level metadata table.
}
\examples{
# Create group-level annotations
scores = data.frame(
  obs_id = 0:3,
  labels = c("T cells", "B cells", "Monocytes", "NK cells")
)
# Observation-level group assignments
clusters = c(0, 0, 1, 2, 1, 3, 0, 2, 1, 3)
# Map labels to observations
obs_labels = expand_clusterlabel(scores, clusters)
print(obs_labels)
}
