% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_by_pred.R
\name{plot_scores_by_pred}
\alias{plot_scores_by_pred}
\title{Boxplot of top scores by predicted label}
\usage{
plot_scores_by_pred(
  scores,
  outlier_shape = 19,
  alpha_val = 0.7,
  jitter = FALSE
)
}
\value{
ggplot object
}
\description{
Boxplot of top scores by predicted label
}
