% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal_vs_noise.R
\name{plot_signal_vs_noise}
\alias{plot_signal_vs_noise}
\title{Signal vs Noise Boxplot}
\usage{
plot_signal_vs_noise(scores, outlier_shape = 19, alpha = 0.7, jitter = FALSE)
}
\value{
ggplot object
}
\description{
Shows the distribution of scores across all candidate labels for each predicted label.
}
