% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_by_pred.R
\name{plot_scores_by_pred}
\alias{plot_scores_by_pred}
\title{Boxplot of top scores by predicted label}
\usage{
plot_scores_by_pred(scores, outlier_shape = 19, alpha = 0.7, jitter = FALSE)
}
\arguments{
\item{scores}{A data.frame that contains columns "obs_id", "labels", and "score" followed by one or more score columns (one for each candidate label)}

\item{outlier_shape}{Integer value indicating the shape of the outlier points}

\item{alpha}{Numeric value indicating the alpha value for the points in the plot}

\item{jitter}{Logical value indicating whether to add geom_jitter to the plot}
}
\value{
ggplot object
}
\description{
Shows the distribution of scores for each predicted label. If you have 3 labels there will be three boxplots drawn in the same plot. Each point is the score of the observations of a label.
}
\examples{
set.seed(15)
m = matrix(rnorm(60), nrow = 20)
colnames(m) = c("TypeA","TypeB","TypeC")
rownames(m) = paste0("c", rep(1:nrow(m)))
pred = add_labels_based_on_max(m)
plot_scores_by_pred(pred, jitter = TRUE)

}
